(maint) fix flaky test and stop lying about :create
Prior to this commit the initial test of `GithubWebhookController.create`
was flaky for two related reasons:
 1. the controller was spawning an additional process (aside from the
one handling the webhook request) to store the Opportunity into the
database,
 2. the test for `create` was testing two distinct things:
   * `201 Created` response from webhook endpoint (potentially a lie,
   * more on that later) an Opportunity was inserted into the database.

Because the controller was spawning an additional process and the test
was checking the side effects of that process, we had a race condition
on our hands. The first hint of the race condition was the
`Process.sleep` in the test. One could certainly extend the sleep time
in the test to avoid the race condition most of the time, merely
delaying the problem and sacrificing test speed, but the more correct
solution is in the implicit answer we give when processing the webhook
payload. Before this commit we were stating that we had definitely, 100%
created the Opportunity in the database, which, as indicated by the
`Process.sleep` was a lie. We *intended* to insert the Opportunity into
the database, but we might not be able to (for a variety of reasons).

This leaves us with two architectural front-runners:
 * a CQRS model, where we *explicitly* state "we _intend_ to write this,
   but we aren't sure if we actually _can_, check back with us later
   using this token"
 * a synchronous write model

Because GitHub doesn't care about what we do with the payload and
whether or not we actually succeed, the CQRS model is a slight misfit.
The synchronous write model, however, provides us with the advantages of
not stating `201 Created` until it's actually in the DB, and if the DB
write *does* fail, our request-process will die with a 500, which *is* a
trigger for GitHub (and Bitbucket) to reattempt payload delivery later,
giving us another chance at inserting it.  Additionally, if/when we have
monitoring, those 500s will be easily surfaced.

Therefore, this commit opts for a synchronous database write during
webhook payload processing and removes the troublesome `Process.sleep`
from the test.