(maint) eliminate Projects.get_or_insert race cond
Prior to this commit `Projects.get_or_insert` contained a race
condition, while partially mitigated by
2df09be3606c1c02da15d825e9a25508758a4cee, is still reachable through a
fairly straightforward interleaving of two issue payloads for the same
net new project arriving at nearly the same time.

```
Issue A             Issue B
   |                   |
create                 |
   |                create
process_issue          |
   |              process_issue
get_or_insert          |
   |               get_or_insert
get -> nil             |
   |               get -> nil
insert                 |
                    insert
```

After discussion on #29, the conclusion was that Projects would be added
manually and _not_ as part of issue/Opportunity processing. With
projects added manually, the issue processing needs to decide what to do
in the case that the project does not yet exist and gracefully handle
that case. This commit opts for logging that an issue came in for a
project we're not tracking, but ultimately discarding the issue and not
attempting to create an Opportunity for it.