increase untracked project log level, refactor test
Prior to this commit we were only logging the payload for an untracked
project at `:info` and the tests were not validating that the log
message was actually sent. This commit ups the log level to `:warn`,
which if we weren't using `capture_log` would pollute the test output.
Since we need to test for the log message anyway, we use `capture_log`
to for the dual purpose of ensuring the message is sent and preventing
it from cluttering the test output. Also since the payload generation is
not germane to the functionality under test, it's been moved to `setup`
functions.