update sha for locked ex_machina dependency
Prior to this commit our ex_machina dependency was pinned to a specific
hash in a github repository; however, the author of that library amended
that commit, orphaning it, which in turn causes `mix deps.get` to fail.
This commit updates the `mix.lock` to the sha of the _new_ commit so
that we can properly retrieve all our dependencies.