(maint) eliminate Projects.get_or_insert race condition (#34)
* (maint) eliminate Projects.get_or_insert race cond

Prior to this commit `Projects.get_or_insert` contained a race
condition, while partially mitigated by
2df09be3606c1c02da15d825e9a25508758a4cee, is still reachable through a
fairly straightforward interleaving of two issue payloads for the same
net new project arriving at nearly the same time.

```
Issue A             Issue B
   |                   |
create                 |
   |                create
process_issue          |
   |              process_issue
get_or_insert          |
   |               get_or_insert
get -> nil             |
   |               get -> nil
insert                 |
                    insert
```

After discussion on #29, the conclusion was that Projects would be added
manually and _not_ as part of issue/Opportunity processing. With
projects added manually, the issue processing needs to decide what to do
in the case that the project does not yet exist and gracefully handle
that case. This commit opts for logging that an issue came in for a
project we're not tracking, but ultimately discarding the issue and not
attempting to create an Opportunity for it.

* increase untracked project log level, refactor test

Prior to this commit we were only logging the payload for an untracked
project at `:info` and the tests were not validating that the log
message was actually sent. This commit ups the log level to `:warn`,
which if we weren't using `capture_log` would pollute the test output.
Since we need to test for the log message anyway, we use `capture_log`
to for the dual purpose of ensuring the message is sent and preventing
it from cluttering the test output. Also since the payload generation is
not germane to the functionality under test, it's been moved to `setup`
functions.

* update sha for locked ex_machina dependency

Prior to this commit our ex_machina dependency was pinned to a specific
hash in a github repository; however, the author of that library amended
that commit, orphaning it, which in turn causes `mix deps.get` to fail.
This commit updates the `mix.lock` to the sha of the _new_ commit so
that we can properly retrieve all our dependencies.