Internationalize Opportunity levels
Prior to this commit Opportunities were storing their associated level
as an English string in the database. This made two immediate things
difficult:

 1. Sorting from easiest to hardest
 2. Migrating data when level names change (beginner => starter)

In addition we have no provisions for being able to internationalize the
application.

This commit decides to use an integer as its internal representation for
the opportunity difficulty level. The first thing necessary task is to
alter the Ecto schema and the database. Because an existing database
might have valid opportunities in the form of "starter", "intermediate",
"advanced", these are migrated to a stringified integer before the table
itself is modified, while explicitly casting from the stringified
integer to an actual integer. Next we modify the edges of our system to
properly modify a payload with "Level:Beginner" into a `1` and similarly
when displaying an opportunity of difficulty level 1 showing `Starter`
by default or `Leicht` if the process locale is German.